requirements.txt
pandas==0.25.1
scipy==1.3.1
tqdm==4.36.1
numpy==1.17.2
tensorflow==1.13.1
six==1.12.0
imageio==2.6.0
h5py==2.9.0
seaborn==0.9.0
matplotlib==3.1.1
Pillow==8.1.0
scikit_learn==0.24.0
torch==1.7.1
torchvision==0.8.2
